Accession_IDs_of_the sequences_analized_by_plmDCA
The datasets contains two subsets that we collected from GISAID repository. The sequences are collected from the beginning of the COVID19 pandemic to the end of October 2021. There are more than 3,5 billion sequences included. Thus the whole  Dataset of the IDs are divided into two sub-sets. The first one for the beginning to the end of May 2021 and the other one for June 2021 to October 2021.